feat(graphql): add "parent queries" for all entities
Add "parent queries" for all entities. This differs slightly from
Diesel-based implementation as it will only get top-level entities when
null is passed for the parent parameters. This is because this is a more
common use-case.. If a "get-all" entities is desired, a new query will
be necessary or these queries will need to be refactored.