feat: add climbers object and simple getter
As I write more of these, the more I think these simple entities
(defined by a tuple-structure with an i32 id as the 0th and only
element) and simple getters (get single column value with a simple type,
where some id column matches the id of the simple entity being queried
for) can all be macro-ized. Obviously, there's some overhead maintaining
this macro, but it'd make it so much easier to add more things or change
the implementation of all the similar operations.