feat: add ascents type and simple getters
This _does not_ add the date window support, yet. This is because there
is not a great native rust-postgres way of getting PostgreSQLs DATERANGE
type. Looks like there's sfackler/rust-postgres-range#18, but it hasn't
been merged in a year and the general state of rust-postgres-range seems
unmaintained.. perhaps @tkbrigham has a good point in
sfackler/rust-postgres#601.. it'd be great if all the PostgreSQL
features were supported in rust-postgres directly.